test(bpv7): move structural-parser tests from checks.rs to tests/parse.rs
checks.rs is the BPSec pipeline-composition suite, but since the
parser restructure it also carried ten tests that exercise only the
public parse API. Move them to tests/parse.rs where the coverage
report has referenced them all along, leaving checks.rs to the
classify/verify/cascade tests its docstring describes. Test bodies
are unchanged apart from replacing the legacy raw_parse_tuple adapter
with direct parse::parse calls.
